proximity: change lock/unlock semantics
The "Pathloss" label was renamed to "Unlock on signal level". This
reflects the new running semantics. The unlock values in slider can be:
"disabled", "weak", "regular" and "good" (these values represent the
thresholds).

It works this way:
1) unlock (threshold) = good, lock = { regular, weak , link-loss }
2) unlock (threshold) = regular, lock = { weak , link-loss }
3) unlock (threshold) = weak, lock = { link-loss }
4) unlock (threshold) = disabled, Nothing to do

In other words, the device is locked when the current signal level is
below than level chosen for unlock (slider value).